requirements.txt
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.3.0
seaborn>=0.11.0
scipy>=1.7.0
scikit-learn>=1.0.0
xgboost>=1.5.0
optuna>=2.10.0
PyYAML>=5.4.0
jupyter>=1.0.0
notebook>=6.4.0
